+++ default.nix
@@ -18,7 +18,7 @@ depOverrides = new: old: {
     ascii-caseless = new.callPackage ./nix/ascii-caseless-0.0.0.0.nix {};
     ascii-numbers = new.callPackage ./nix/ascii-numbers-1.1.0.2.nix {};
     ascii-predicates = new.callPackage ./nix/ascii-predicates-1.0.1.1.nix {};
-    ascii-superset = new.callPackage ./nix/ascii-superset-1.1.0.0.nix {};
+    ascii-superset = new.callPackage ./nix/ascii-superset-1.2.0.0.nix {};
     ascii-th = new.callPackage ./nix/ascii-th-1.1.1.0.nix {};
 };