+++ nix/ascii-numbers-1.1.0.2.nix
@@ -5,8 +5,8 @@ mkDerivation {
   pname = "ascii-numbers";
   version = "1.1.0.2";
   sha256 = "dbfe4acb0eaa2f0c4ce68c6ac6072297cc485b5c7f9b3ad59375c36133b61837";
-  revision = "1";
-  editedCabalFile = "1g3aazd8jdadzhf9dp3p1x5m3bdvng0zmg4madkg9750qh0vdspr";
+  revision = "2";
+  editedCabalFile = "19x9mh11pb7j4ykf9vicprn6mlhcb9gwsk82gh5yk366k4r172d7";
   libraryHaskellDepends = [
     ascii-case ascii-char ascii-superset base bytestring hashable text
   ];